+++ apollo/src/resolvers/index.js
@@ -1,5 +1,5 @@
 const Booking = require('./Booking')
-const Mutation = require('./Mutation')
+const Mutation = require('./mutation')
 const Query = require('./Query')
 const Response = require('./Response')
 const ResumeReview = require('./ResumeReview')
@@ -9,12 +9,12 @@ const User = require('./User')
 
 const resolvers = {
   // Booking,
-  // Mutation,
+  Mutation,
   // Query,
   // Response,
   // ResumeReview,
   // Review,
-  User
+  // User
 }
