+++ apollo/src/resolvers/mutation/index.js
@@ -1,0 +1,11 @@
+const { Stripe } = require('./Stripe')
+// const Mutation = require('./Mutation')
+const { User } = require('./User')
+
+//  Breaks up and imports Mutation resolvers by concern
+
+
+const Mutation = { ...Stripe, ...User }
+
+
+module.exports = Mutation
